move load_network_for_plots to plot_network
