+++ scripts/plot_network.py
@@ -23,6 +23,7 @@ import matplotlib as mpl
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import pypsa
 from matplotlib.legend_handler import HandlerPatch
 from matplotlib.patches import Circle, Ellipse
 
@@ -32,9 +33,9 @@ from scripts._helpers import (
     change_to_script_dir,
     configure_logging,
     create_logger,
-    load_network_for_plots,
     mock_snakemake,
 )
+from scripts.add_electricity import load_costs, update_transmission_costs
 
 to_rgba = mpl.colors.colorConverter.to_rgba
 
@@ -355,6 +356,40 @@ def plot_total_cost_bar(n, ax=None):
     ax.set_xlim([0, 1])
     ax.set_xticklabels([])
     ax.grid(True, axis="y", color="k", linestyle="dotted")
+
+
+def load_network_for_plots(
+    fn, tech_costs, cost_config, elec_config, combine_hydro_ps=True
+):
+
+    n = pypsa.Network(fn)
+
+    n.loads["carrier"] = n.loads.bus.map(n.buses.carrier) + " load"
+    n.stores["carrier"] = n.stores.bus.map(n.buses.carrier)
+
+    n.links["carrier"] = (
+        n.links.bus0.map(n.buses.carrier) + "-" + n.links.bus1.map(n.buses.carrier)
+    )
+    n.lines["carrier"] = "AC line"
+    n.transformers["carrier"] = "AC transformer"
+
+    n.lines["s_nom"] = n.lines["s_nom_min"]
+    n.links["p_nom"] = n.links["p_nom_min"]
+
+    if combine_hydro_ps:
+        n.storage_units.loc[
+            n.storage_units.carrier.isin({"PHS", "hydro"}), "carrier"
+        ] = "hydro+PHS"
+
+    # if the carrier was not set on the heat storage units
+    # bus_carrier = n.storage_units.bus.map(n.buses.carrier)
+    # n.storage_units.loc[bus_carrier == "heat","carrier"] = "water tanks"
+
+    Nyears = n.snapshot_weightings.objective.sum() / 8760.0
+    costs = load_costs(tech_costs, cost_config, elec_config, Nyears)
+    update_transmission_costs(n, costs)
+
+    return n
 
 
 if __name__ == "__main__":
